requirements.txt
astroid==3.1.0
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
fonttools==4.51.0
isort==5.13.2
kiwisolver==1.4.5
matplotlib==3.8.4
mccabe==0.7.0
numpy==1.26.4
opencv-contrib-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.0
pylint==3.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tomlkit==0.12.4
tzdata==2024.1
